[Misc Utils] New module
