+++ deepchem/utils/differentiation_utils/misc.py
@@ -131,43 +131,3 @@ def assert_runtime(cond, msg=""):
     """
     if not cond:
         raise RuntimeError(msg)
-
-
-# repr helper functions
-def indent(s, nspace):
-    """give indentation of the second line and next lines
-
-    Parameters
-    ----------
-    s: str
-        The string to be indented.
-    nspace: int
-        The number of spaces to be indented.
-
-    Returns
-    -------
-    str
-        The indented string.
-
-    """
-    spaces = " " * nspace
-    lines = [spaces + c if i > 0 else c for i, c in enumerate(s.split("\n"))]
-    return "\n".join(lines)
-
-
-def shape2str(shape):
-    """Convert the shape to string representation.
-    It also nicely formats the shape to be readable.
-
-    Parameters
-    ----------
-    shape: Sequence[int]
-        The shape to be converted to string representation.
-
-    Returns
-    -------
-    str
-        The string representation of the shape.
-
-    """
-    return "(%s)" % (", ".join([str(s) for s in shape]))
